refactor(sdk): split public API into multiple files
